Test shipping address presentation state persistence
Committed-By-Agent: codex
Committed-By-Agent: codex
Committed-By-Agent: codex
